Add mod prune rule
`2 * x % x` is always 0 as `x = x`. So are all expressions `x^p % x^q`
where `p >= q`, as this is equal to `x^(p-q) * x^q % x^q`, which when
substituting `y = x^q` is `x^(p-q) * y % y` and `y = y`.

This is an adaptation of the `divToPowSub` `Prune` rule of the `div` op.